feat: #566 add isProgress field to the way model (#570)

+++ mw-server/schemas/jobDones.schema.go
@@ -6,18 +6,18 @@ import (
 
 type CreateJobDonePayload struct {
 	Description   string    `json:"description"`
-	Time          int64     `json:"time"`
+	Time          int32     `json:"time"`
 	DayReportUuid uuid.UUID `json:"dayReportUuid"`
 	OwnerUuid     uuid.UUID `json:"ownerUuid"`
 }
 
 type UpdateJobDone struct {
 	Description string `json:"description"`
-	Time        int64  `json:"time"`
+	Time        int32  `json:"time"`
 }
 
 type JobDonePlainResponse struct {
 	Description string            `json:"description"`
-	Time        int64             `json:"time"`
+	Time        int32             `json:"time"`
 	Owner       UserPlainResponse `json:"owner"`
 }
